requirements.txt
Django==5.0.1
djangorestframework==3.15.0
PyMySQL==1.1.0
djangorestframework-simplejwt==5.3.1
...your existing packages...
selenium==4.18.1
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
lxml==5.1.0